generic computer player
Update computer player to rely only on the functions specified in the game protocol.

Also updated the logic around turn taking and marker selection to take advantage of
the functions provided by the game protocol.

Added option arguments to allow configuration for the behavior of the perfect move
computer player function.