fix resignation logic fix timeout after opp resign
resignation logic was firing off the resign function twice, now terminate is the only function that will send a resign message to the game server.

timeout logic was causing a LiveView crash after the opp resign if the winning player stayed in the liveview for longer than the max turn length.  Updated turn logic so that the timer will stop if the liveview is in the opp-resigned state.